feed/mirror-tv/google_news_rss/requirements.txt
astroid==2.4.2
autopep8==1.5.4
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
DateTime==4.3
feedgen==0.9.0
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.3
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
gql==3.0.0a5
graphql-core==3.1.2
idna==2.10
isort==5.5.2
lazy-object-proxy==1.4.3
lxml==4.6.1
mccabe==0.6.1
multidict==5.0.2
promise==2.3
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
rsa==4.6
Rx==1.6.1
six==1.15.0
toml==0.10.1
urllib3==1.26.2
wrapt==1.12.1
yarl==1.6.3
zope.interface==5.2.0
